empleados

+++ src/main/java/edu/pe/demale/demale_proyecto/repositories/PersonaRepository.java
@@ -1,4 +1,4 @@
-package edu.pe.demale.demale_proyecto.repositories;
+package edu.pe.demale.demale_proyecto.repositories; // Confirma que tu paquete es 'repositories' (plural)
 
 import edu.pe.demale.demale_proyecto.models.Persona;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface PersonaRepository extends JpaRepository<Persona, Long> {
+public interface PersonaRepository extends JpaRepository<Persona, Integer> { // <-- CAMBIO AQUÍ: Long a Integer
     Optional<Persona> findByNumeroDocumento(String numeroDocumento);
 }
